chore: clean up and add test cases
